requirements.txt
Flask==2.3.2
vonage==2.9.0
google-generativeai==0.19.0
